server/requirements.txt
fastapi
uvicorn[standard]
spotipy
yt-dlp
python-dotenv
ffmpeg-python